% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{A3SS.info.seperation}
\alias{A3SS.info.seperation}
\title{Split A3SS event information}
\usage{
A3SS.info.seperation(event)
}
\arguments{
\item{event}{a event id}
}
\value{
data.frame
}
\description{
Split A3SS event information
}
