% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{A3SS.kmer.extraction}
\alias{A3SS.kmer.extraction}
\title{get kmer of exons and splice region associated with A3SS event}
\usage{
A3SS.kmer.extraction(splice.region, exon.region)
}
\arguments{
\item{splice.region}{data.frame with sequence of splice region, \code{A3SS.splice.seq.extraction}}

\item{exon.region}{data.frame with sequence of exon region, \code{A3SS.exon.seq.extraction}}
}
\value{
vector of sequence kmer
}
\description{
get kmer of exons and splice region associated with A3SS event
}
\keyword{internal}
