% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{A3SS.length.feature}
\alias{A3SS.length.feature}
\title{get length of exons and introns associated with A3SS event}
\usage{
A3SS.length.feature(event)
}
\arguments{
\item{event}{a event id}
}
\value{
data.frame
}
\description{
get length of exons and introns associated with A3SS event
}
\keyword{internal}
