% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{A5SS.info.seperation}
\alias{A5SS.info.seperation}
\title{Split A5SS event information}
\usage{
A5SS.info.seperation(event)
}
\arguments{
\item{event}{a event id}
}
\value{
data.frame
}
\description{
Split A5SS event information
}
