% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{A5SS.splice.seq.extraction}
\alias{A5SS.splice.seq.extraction}
\title{Extract splice sequence for A5SS event}
\usage{
A5SS.splice.seq.extraction(events.info, bs.genome, core)
}
\arguments{
\item{events.info}{data.frame A5SS event information}

\item{bs.genome}{\code{get(pkg)}, see \code{createBSgenome} function}

\item{core}{the number of threads}
}
\value{
list
}
\description{
Extract splice sequence for A5SS event
}
\keyword{internal}
