% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calcu_ex_rc}
\alias{calcu_ex_rc}
\title{Calculate the mean of exclusion read counts}
\usage{
calcu_ex_rc(rc, event)
}
\arguments{
\item{rc}{matrix, reads associated with splicing events}

\item{event}{data frame of event information}
}
\value{
matrix of the mean of exclusion read counts
}
\description{
exclusion reads recorded in event
}
\keyword{internal}
