% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calcu_in_rc}
\alias{calcu_in_rc}
\title{Calculate the mean of inclusion read counts}
\usage{
calcu_in_rc(rc, event)
}
\arguments{
\item{rc}{matrix, reads associated with splicing events}

\item{event}{data frame of event information}
}
\value{
matrix of the mean of inclusion read counts
}
\description{
inclusion reads recorded in event
}
\keyword{internal}
