% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calcu_rc}
\alias{calcu_rc}
\title{Calculate the mean of read counts}
\usage{
calcu_rc(rc, event)
}
\arguments{
\item{rc}{matrix, reads associated with splicing events}

\item{event}{data frame of event information}
}
\value{
matrix of the mean of read counts
}
\description{
associated reads name recorded in event
}
\keyword{internal}
