% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimilarity.R
\name{createBSgenome}
\alias{createBSgenome}
\title{Create R package for a reference sequence file}
\usage{
createBSgenome(ref_path, out_path, pkg, install_lib = NULL)
}
\arguments{
\item{ref_path}{reference fasta file}

\item{out_path}{path to save the package}

\item{pkg}{the name of pkg}

\item{install_lib}{library path that save the package}
}
\value{
package path
}
\description{
BSgenome used for sequence extraction
}
