% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{filterMXE}
\alias{filterMXE}
\title{MXE Filter}
\usage{
filterMXE(event, rc, min.Cell)
}
\arguments{
\item{event}{data frame of event information}

\item{rc}{data frame of reads associated with splicing events}

\item{min.Cell}{the minimum number of cells that are required to have at least
one reads of both upstream and downstream junction supporting exon}
}
\value{
data frame of filtered event information
}
\description{
MXE need strict filter of upstream and downstream junction reads
}
