% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneExpression.R
\name{get10XEXPmatrix}
\alias{get10XEXPmatrix}
\title{Get gene expression matrix from Cell Ranger}
\usage{
get10XEXPmatrix(
  paras,
  expr_path,
  filter.mt = paras$Basic$filter_sc$filter.mt,
  filter.rp = paras$Basic$filter_sc$filter.rp
)
}
\arguments{
\item{paras}{list fromJSON(paras_file)
Default dataset, filter.mt, filter.rp from paras}

\item{expr_path}{directory to the cellranger output}

\item{filter.mt}{filter out mitochondrial genes}

\item{filter.rp}{filter out ribosomal genes}
}
\value{
gene expression matrix path
}
\description{
Read count matrix from 10X CellRanger hdf5 file
save raw and normalized UMI counts sparse matrix to work_path/rds/
}
