% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEvent.R
\name{getA35SSALevent}
\alias{getA35SSALevent}
\title{Detect A3SS,A5SS,AL events}
\usage{
getA35SSALevent(
  work_path,
  bam_path,
  gff,
  script_majiq,
  core,
  genome_name,
  junctionReads,
  MAJIQ_env,
  log_file
)
}
\arguments{
\item{work_path}{directory to deposite all SCSES results in}

\item{bam_path}{directory to Pseudobulk bam file}

\item{gff}{the gene annotation in gff format}

\item{script_majiq}{directory to script for running MAJIQ}

\item{core}{the number of threads}

\item{genome_name}{Genome name required by MAJIQ}

\item{junctionReads}{the minimum total number of reads for any junction}

\item{MAJIQ_env}{the name of conda env of MAJIQ}

\item{log_file}{file saving stdout and stderr information}
}
\value{
Splicing event directory
}
\description{
Detect A3SS,A5SS,AL events
}
