% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEvent.R
\name{getA3SSid}
\alias{getA3SSid}
\title{Generate A3SS event id}
\usage{
getA3SSid(file, outfile, remove.chr = F, junctionReads)
}
\arguments{
\item{file}{directory to MAJIQ output}

\item{outfile}{directory to output file A3SS.txt}

\item{remove.chr}{if TRUE,remove the "chr" character of event id}

\item{junctionReads}{the minimum total number of reads for any junction}
}
\value{
The number of A3SS events
}
\description{
Generate A3SS event id
}
