% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEvent.R
\name{getALid}
\alias{getALid}
\title{Generate AL event id}
\usage{
getALid(file, outfile, remove.chr = F, junctionReads)
}
\arguments{
\item{file}{directory to MAJIQ output}

\item{outfile}{directory to output file AL.txt}

\item{remove.chr}{if TRUE,remove the "chr" character of event id}

\item{junctionReads}{the minimum total number of reads for any junction}
}
\value{
The number of AL events
}
\description{
Generate AL event id
}
