% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEvent.R
\name{getEvent}
\alias{getEvent}
\title{Detect a specific type of splicing events}
\usage{
getEvent(paras, event_type)
}
\arguments{
\item{paras}{list, read json configuration files}

\item{event_type}{splicing event type (A3SS,A5SS,AL,SE,MXE,RI)}
}
\value{
Splicing event directory
}
\description{
Detect a specific type of splicing events based on Pseudobulk
bam in work_path/data and generate event id saved to work_path/events
}
