% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneExpression.R
\name{getEXPmatrix}
\alias{getEXPmatrix}
\title{Get gene expression matrix}
\usage{
getEXPmatrix(
  paras,
  expr_path = NULL,
  dataset = paras$DataSet,
  filter.mt = paras$Basic$filter_sc$filter.mt,
  filter.rp = paras$Basic$filter_sc$filter.rp
)
}
\arguments{
\item{paras}{list fromJSON(paras_file)
Default dataset, filter.mt, filter.rp from paras}

\item{expr_path}{directory to the featureCounts output}

\item{dataset}{the name of dataset}

\item{filter.mt}{filter out mitochondrial genes}

\item{filter.rp}{filter out ribosomal genes}
}
\value{
gene expression matrix path
}
\description{
save gene expression count and TPM matrix to work_path/rds/
}
