% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEvent.R
\name{getMXEid}
\alias{getMXEid}
\title{Generate MXE event id}
\usage{
getMXEid(file, outfile, remove.chr = F, junctionReads)
}
\arguments{
\item{file}{directory to rMats output}

\item{outfile}{directory to output file MXE.txt}

\item{remove.chr}{if TRUE,remove the "chr" character of event id}

\item{junctionReads}{the minimum total number of reads for any junction}
}
\value{
The number of MXE events
}
\description{
Generate MXE event id
}
