% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{getPredictProb}
\alias{getPredictProb}
\title{Predict the probabilities of WD, BD, and TD-Info for each event-cell pair.
first: probabilities of being WD but not ND for each event-cell pair;
second: probabilities of being BD but not TD for each event-cell pair;
third: probabilities of being TD+Info but not TD-Info for each event-cell pair.}
\usage{
getPredictProb(feature_df, model1, model2, psi)
}
\arguments{
\item{feature_df}{a data.frame of features for classifer}

\item{model1}{logistical model to distinguish BD and TD}

\item{model2}{logistical model to distinguish TD+Info and TD-Info}

\item{psi}{matrix of psi value before imputation}
}
\value{
a list of probability matrix
}
\description{
Predict the probabilities of WD, BD, and TD-Info for each event-cell pair.
first: probabilities of being WD but not ND for each event-cell pair;
second: probabilities of being BD but not TD for each event-cell pair;
third: probabilities of being TD+Info but not TD-Info for each event-cell pair.
}
