% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEvent.R
\name{getSEevent}
\alias{getSEevent}
\title{Detect SE events}
\usage{
getSEevent(
  work_path,
  bam_path,
  gtf,
  paired,
  readlength,
  rmats_path,
  py_path,
  script_rmats,
  core,
  log_file
)
}
\arguments{
\item{work_path}{directory to deposite all SCSES results in}

\item{bam_path}{directory to Pseudobulk bam file}

\item{gtf}{the gene annotation in gtf format}

\item{paired}{type of read: either "paired" for paired-end  or "single" for single-end}

\item{readlength}{the length of each read}

\item{rmats_path}{directory to rmats.py}

\item{py_path}{directory to python for running rMats}

\item{script_rmats}{directory to script for running rMats}

\item{core}{the number of threads}

\item{log_file}{file saving stdout and stderr information}
}
\value{
Splicing event directory
}
\description{
Detect SE events
}
