% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeFtSplicingValue.R
\name{mergeFtSplicingValue}
\alias{mergeFtSplicingValue}
\title{Merge splicing events for classifer fine tune}
\usage{
mergeFtSplicingValue(paras)
}
\arguments{
\item{paras}{list readSCSESconfig(paras_file)}
}
\value{
Merged splicing value path
}
\description{
Merge different types of splicing events in
work_path/splicing_value_ft, and save to work_path/rds_ft
}
