% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeSplicingValue.R
\name{mergeSplicingValue}
\alias{mergeSplicingValue}
\title{Merge splicing events}
\usage{
mergeSplicingValue(paras)
}
\arguments{
\item{paras}{list fromJSON(paras_file)}
}
\value{
Merged splicing value path
}
\description{
Merge different types of splicing events
in work_path/splicing_value, and save merged splicing value to work_path/rds
}
