% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeSplicingValue.R
\name{mgergeEvent}
\alias{mgergeEvent}
\title{Merge event information}
\usage{
mgergeEvent(splicing_value_path, output_path)
}
\arguments{
\item{splicing_value_path}{The path to *_rc.rds
* represents different types of splicing events}

\item{output_path}{The path to save merged event information}
}
\value{
Merged event information path
}
\description{
Merge event information for different types of splicing events
in splicing_value_path, and save merged event information to output_path
}
