% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeSplicingValue.R
\name{mgergeRC}
\alias{mgergeRC}
\title{Merge Read counts}
\usage{
mgergeRC(splicing_value_path, output_path)
}
\arguments{
\item{splicing_value_path}{The path to *_rc.rds
* represents different types of splicing events}

\item{output_path}{The path to save merged Read counts}
}
\value{
Merged Read counts path
}
\description{
Merge Read counts for different types of splicing events
in splicing_value_path, and save merged Read counts to output_path
}
