% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{MXE.info.seperation}
\alias{MXE.info.seperation}
\title{Split MXE event information}
\usage{
MXE.info.seperation(event)
}
\arguments{
\item{event}{a event id}
}
\value{
data.frame
}
\description{
Split MXE event information
}
\keyword{internal}
