% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{MXE.length.feature}
\alias{MXE.length.feature}
\title{get length of exons and introns associated with MXE event}
\usage{
MXE.length.feature(event)
}
\arguments{
\item{event}{a event id}
}
\value{
data.frame
}
\description{
get length of exons and introns associated with MXE event
}
\keyword{internal}
