% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimilarity.R
\name{PCA_D_reduct}
\alias{PCA_D_reduct}
\title{PCA}
\usage{
PCA_D_reduct(data, id_select)
}
\arguments{
\item{data}{a matrix or data frame}

\item{id_select}{used features}
}
\value{
principal components retain ~90% of the variation in the data
}
\description{
Perform a principal components analysis on the given data 
and return the results
}
