% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimilarity.R
\name{PCA_D_reduct_sparse}
\alias{PCA_D_reduct_sparse}
\title{PCA for sparse matrix}
\usage{
PCA_D_reduct_sparse(data, id_select)
}
\arguments{
\item{data}{a sparse matrix}

\item{id_select}{used features}
}
\value{
principal components retain ~90% of the variation in the data
}
\description{
Perform a principal components analysis on the given sparse 
matrix and return the results
}
