% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{phastScore.extraction}
\alias{phastScore.extraction}
\title{extract conservation scores of genome region}
\usage{
phastScore.extraction(regionlist, bwf, chr.prefix, core)
}
\arguments{
\item{regionlist}{list of genome region}

\item{bwf}{\code{BigWigFile(phast.path)},
phast.path can be downloaded from ucsc database,
e.g. https://hgdownload.cse.ucsc.edu/goldenpath/hg19/phastCons100way/}

\item{chr.prefix}{prefix added to chromosome id, "chr" or ""}

\item{core}{the number of threads}
}
\value{
list
}
\description{
extract conservation scores of genome region
}
\keyword{internal}
