% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rc_to_psi}
\alias{rc_to_psi}
\title{Calculate psi}
\usage{
rc_to_psi(event, rc)
}
\arguments{
\item{event}{data frame of event information,
event types: A3SS, A5SS, AL, SE, MXE, RI}

\item{rc}{matrix, reads associated with splicing events}
}
\value{
data.frame of psi value
}
\description{
Calculate psi
}
\keyword{internal}
