% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createConfig.R
\name{readSCSESconfig}
\alias{readSCSESconfig}
\title{read configure}
\usage{
readSCSESconfig(paras_file)
}
\arguments{
\item{paras_file}{path to configure file}
}
\value{
list of parameters
}
\description{
read configure
}
