% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{RI.info.seperation}
\alias{RI.info.seperation}
\title{Split RI event information}
\usage{
RI.info.seperation(event)
}
\arguments{
\item{event}{a event id}
}
\value{
data.frame
}
\description{
Split RI event information
}
