% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{RI.kmer.extraction}
\alias{RI.kmer.extraction}
\title{get kmer of exons and splice region associated with RI event}
\usage{
RI.kmer.extraction(splice.region, exon.region)
}
\arguments{
\item{splice.region}{data.frame with sequence of splice region, \code{RI.splice.seq.extraction}}

\item{exon.region}{data.frame with sequence of exon region, \code{RI.exon.seq.extraction}}
}
\value{
vector of sequence kmer
}
\description{
get kmer of exons and splice region associated with RI event
}
\keyword{internal}
