% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{SE.exon.seq.extraction}
\alias{SE.exon.seq.extraction}
\title{Extract exon sequence for SE event}
\usage{
SE.exon.seq.extraction(events.info, bs.genome, core)
}
\arguments{
\item{events.info}{data.frame SE event information}

\item{bs.genome}{\code{get(pkg)}, see \code{createBSgenome} function}

\item{core}{the number of threads}
}
\value{
list
}
\description{
Extract exon sequence for SE event
}
\keyword{internal}
