% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{SE.info.seperation}
\alias{SE.info.seperation}
\title{Split SE event information}
\usage{
SE.info.seperation(event)
}
\arguments{
\item{event}{a event id}
}
\value{
data.frame
}
\description{
Split SE event information
}
\keyword{internal}
