% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFeatureFunction.R
\name{SE.length.feature}
\alias{SE.length.feature}
\title{get length of exons and introns associated with SE event}
\usage{
SE.length.feature(event)
}
\arguments{
\item{event}{a event id}
}
\value{
data.frame
}
\description{
get length of exons and introns associated with SE event
}
\keyword{internal}
