% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createConfig.R
\name{createConfigshiny}
\alias{createConfigshiny}
\title{Create configure file}
\usage{
createConfigshiny(host = "127.0.0.1", port = 9999, launch.browser = F, ...)
}
\arguments{
\item{host}{The IPv4 address that the application should listen on, default: 127.0.0.1}

\item{port}{The TCP port that the application should listen on, default: 9999}

\item{launch.browser}{If true, the system's default web browser will be launched automatically after the app is started.}
}
\description{
Create configure file
}
