% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPseudobulk.R
\name{createPseudobulk}
\alias{createPseudobulk}
\title{Create Pseudobulk bam by merging single cell bam files and index}
\usage{
createPseudobulk(
  paras,
  bam_path = paras$Basic$bam_path,
  core = paras$Basic$core
)
}
\arguments{
\item{paras}{list fromJSON(paras_file)
Default bam_path, samtools_path, core from paras}

\item{bam_path}{directory to single cell bam file}

\item{core}{the number of threads}
}
\value{
Pseudobulk path
}
\description{
Create Pseudobulk bam by merging single cell bam files and index
}
