% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEvent.R
\name{detectEvents}
\alias{detectEvents}
\title{Detect splicing events and generate event id}
\usage{
detectEvents(paras)
}
\arguments{
\item{paras}{list fromJSON(paras_file)}
}
\value{
Splicing event directory
}
\description{
Detect splicing events based on Pseudobulk bam
in work_path/data and generate event id saved to work_path/events
}
