% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{FtClassifier}
\alias{FtClassifier}
\title{Fine-tune the classifer determining the groups that event-cell pairs belong to}
\usage{
FtClassifier(
  paras,
  rds_path = NULL,
  rds_ft_path = NULL,
  rds_cell_similarity_path = NULL,
  decay_impute = paras$Task$impute$decay_impute
)
}
\arguments{
\item{paras}{list fromJSON(paras_file)
Default decay_impute from paras}

\item{rds_path}{path to processed rds data}

\item{rds_ft_path}{path to rds data of fine tune splicing events}

\item{rds_cell_similarity_path}{path to cell similarity, list with name(EXP_RBP, RC and PSI)}

\item{decay_impute}{threshold of change in the similarity matrix}
}
\value{
path to final classifer
}
\description{
Model1 is used to predict the probabilities belonging to BD and TD,
and Model2 is used to predict probabilities belonging to TD-Info and TD+Info
}
