% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimilarity.R
\name{getCellSimilarity}
\alias{getCellSimilarity}
\title{KNN graph construction for cells}
\usage{
getCellSimilarity(
  paras,
  rds_path = NULL,
  output_path = NULL,
  feature_num = paras$Task$impute$feature_num,
  rbp = paras$Task$impute$rbp,
  cell_similarity_data = paras$Task$impute$cell_similarity_data,
  distance_method = paras$Task$impute$KNN$cell$distance_method,
  alpha_cell = paras$Task$impute$KNN$cell$alpha,
  decay_cell = paras$Task$impute$KNN$cell$decay,
  kcell_max = paras$Task$impute$KNN$cell$kmax,
  kcell_min = paras$KNN$cell$kmin,
  cell.select = NULL
)
}
\arguments{
\item{paras}{list fromJSON(paras_file)
Default core, pkg, ref_path,phast.path,chr.prefix, ae.para,
rbp, kevent, alpha_event,decay_event from paras}

\item{rds_path}{path to processed rds data}

\item{output_path}{path to event similarity}

\item{feature_num}{the number of high variable features for PCA}

\item{rbp}{path to rbp or character string of rbp}

\item{cell_similarity_data}{data used to calculate cell similarity.
Choose at least on from EXP_RBP, RC, and PSI, seperated by ";",
EXP_RBP means TPM/normalized UMI counts for rbp, default: extract from rds_processed/expr.rds
RC means read counts associated with splicing events,  default: rds_processed/rc.rds
PSI meand psi value of splicing events,  default: rds_processed/psi.rds}

\item{distance_method}{method used to calculate distance,
euclidean or cosine}

\item{alpha_cell}{restart probability for random walk}

\item{decay_cell}{threshold of change in the similarity matrix}

\item{kcell_max}{Maximum number of neighbors}

\item{kcell_min}{Minimum number of neighbors}

\item{cell.select}{cells selected for analysis, should match the name
of single-cell bam file name excluding the .bam suffix}
}
\value{
path to cell similarity
save a list of different types of cell similarity and a list of the number 
of neighbors to rds file to work_path/imputation/cell_similarity/cell.similars.rds 
and work_path/imputation/cell_similarity/dyk.cell.rds
}
\description{
Computes cell distance and the k nearest
neighbors for given dataset by data types.
}
