% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimilarity.R
\name{getEventSimilarity}
\alias{getEventSimilarity}
\title{KNN graph construction for events}
\usage{
getEventSimilarity(
  paras,
  rds_path = NULL,
  output_path = NULL,
  core = paras$Basic$core,
  pkg = paras$Basic$refgenome$genome_name,
  ref_path = paras$Basic$refgenome$ref_path,
  phast.path = paras$Task$impute$event_features$phast_path,
  chr.prefix = paras$Task$impute$event_features$chr_prefix,
  ae.para = paras$Task$impute$event_features$AE,
  rbp = paras$Task$impute$rbp,
  kevent = paras$Task$impute$KNN$event$k,
  alpha_event = paras$Task$impute$KNN$event$alpha,
  decay_event = paras$Task$impute$KNN$event$decay
)
}
\arguments{
\item{paras}{list readSCSESconfig(paras_file)
Default core, pkg, ref_path,phast.path,chr.prefix, ae.para,
rbp, kevent, alpha_event,decay_event from paras}

\item{rds_path}{path to processed rds data}

\item{output_path}{path to event similarity}

\item{core}{the number of threads}

\item{pkg}{the name of genome}

\item{ref_path}{path to fasta file}

\item{phast.path}{path to phastCons.bw file}

\item{chr.prefix}{prefix added to chromosome id, "chr" or ""}

\item{ae.para}{parameters of encoding sequence features}

\item{rbp}{path to rbp or character string of rbp}

\item{kevent}{the number of neighbors}

\item{alpha_event}{1-restart probability for random walk}

\item{decay_event}{threshold of change in the similarity matrix}
}
\value{
path to event similarity
save a list of event similarity for different types of splicing events to
work_path/imputation/event_similarity/event.similars.rds
}
\description{
Computes splicing event features and the k nearest
neighbors for given dataset by event types
}
