% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFtSplicingValue.R
\name{getFtRawPSI}
\alias{getFtRawPSI}
\title{Calculate PSI value of splicing events for classifer fine tune}
\usage{
getFtRawPSI(paras)
}
\arguments{
\item{paras}{list readSCSESconfig(paras_file)}
}
\value{
raw PSI path
}
\description{
Save raw PSI for events in ft_event_path
save to work_path/splicing_value_ft/*_psi.rds
}
