% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFtSplicingValue.R
\name{getFtRawRC}
\alias{getFtRawRC}
\title{Quantify raw reads of splicing events for classifer fine tune}
\usage{
getFtRawRC(
  paras,
  bam_path = paras$Basic$bam_path,
  core = paras$Basic$core,
  sequence = paras$Basic$sequence,
  genome_name = paras$Basic$refgenome$genome_name
)
}
\arguments{
\item{paras}{list fromJSON(paras_file)
Default bam_path, java_path, core, and sequence from paras}

\item{bam_path}{The path to save single cell bam file}

\item{core}{the number of threads}

\item{sequence}{full_length or UMI}

\item{genome_name}{Genome name, hg19 or hg38}
}
\value{
raw read path
}
\description{
Quantify raw reads for events in ft_event_path
save to work_path/splicing_value_ft/*_rc.rds
}
