% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneExpression.R
\name{getGeneExpression}
\alias{getGeneExpression}
\title{Get gene expression by featureCounts}
\usage{
getGeneExpression(
  paras,
  bam_path = paras$Basic$bam_path,
  gtf = paras$Basic$refgenome$gtf_path,
  ref = paras$Basic$refgenome$ref_path,
  dataset = paras$DataSet,
  pair = paras$Basic$paired,
  core = paras$Basic$core
)
}
\arguments{
\item{paras}{list readSCSESconfig(paras_file)
Default bam_path, gtf, ref, script, featureCounts_path,
dataset, and core from paras}

\item{bam_path}{directory to single cell bam file}

\item{gtf}{the gene annotation in gtf format}

\item{ref}{fasta file}

\item{dataset}{the name of dataset}

\item{pair}{type of read: either "paired" for paired-end or "single" for single-end}

\item{core}{the number of threads}
}
\value{
featureCounts output path
}
\description{
save featureCounts output to work_path/expr
}
