% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawSplicingValue.R
\name{getRawPSI}
\alias{getRawPSI}
\title{Calculate PSI value for different types of splicing events}
\usage{
getRawPSI(paras, event_types = NULL)
}
\arguments{
\item{paras}{list readSCSESconfig(paras_file)}

\item{event_types}{splicing event types seperated by ";",default=NULL}
}
\value{
raw PSI path
}
\description{
Save raw PSI for events in work_path/events
save to work_path/splicing_value/*_psi.rds
}
