% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawSplicingValue.R
\name{getRawRC}
\alias{getRawRC}
\title{Quantify raw reads for different types of splicing events}
\usage{
getRawRC(
  paras,
  bam_path = paras$Basic$bam_path,
  core = paras$Basic$core,
  sequence = paras$Basic$sequence,
  event_types = NULL
)
}
\arguments{
\item{paras}{list fromJSON(paras_file)
Default bam_path, java_path, core, and sequence from paras}

\item{bam_path}{The path to save single cell bam file}

\item{core}{the number of threads}

\item{sequence}{full-length or UMI}

\item{event_types}{splicing event types seperated by ";",default=NULL}
}
\value{
raw read path
}
\description{
Quantify raw reads for events in work_path/events
save to work_path/splicing_value/*_rc.rds
}
