% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{Imputation}
\alias{Imputation}
\title{Perform SCSES imputation}
\usage{
Imputation(
  paras,
  output_path = NULL,
  decay_impute = paras$Task$impute$decay_impute,
  rc = NULL,
  psi = NULL,
  event.info = NULL,
  cell_similarity = NULL,
  event_similarity = NULL,
  cell_similarity_type
)
}
\arguments{
\item{paras}{list fromJSON(paras_file)
Default decay_impute from paras}

\item{output_path}{path to event similarity}

\item{decay_impute}{threshold of change in the similarity matrix}

\item{rc}{matrix of normalized splicing events associated read counts or path
to a rds file, default: rds_processed/rc.rds}

\item{psi}{matrix of psi value or path to a rds file,
default: rds_processed/psi.rds}

\item{event.info}{list of events information or path to a rds file,
default: rds_processed/event.info.list.rds}

\item{cell_similarity}{a list of different types of cell similarity named by the data
type (EXP_RBP, RC, and PSI) or path to a rds file,
default: imputation/cell_similarity/cell.similars.rds}

\item{event_similarity}{a list of event similarity named by different types of
splicing events or path to a rds file,
default: imputation/event_similarity/event.similars.rds}

\item{cell_similarity_type}{one data type used to calculate cell similarity (EXP_RBP, RC, PSI)}
}
\value{
path to the result of three imputation strategies
save a list of result of three imputation strategies to
work_path/imputation/Imputed_seperated_event_${cell_similarity_type}_*.rds
The result contains two lists, each of which contains multiple imputation results.
The first list is the result of strategy1 and strategy2,
The second list is the result of strategy3
}
\description{
run three PSI imputation strategies
for a sepcific type of cell similarity
}
