% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocessEvent}
\alias{preprocessEvent}
\title{Get Filtered gene expression and splicing value data}
\usage{
preprocessEvent(
  paras,
  min.percentCells.gene = paras$Basic$filter_sc$min.percentCells.gene,
  min.percentCells.event = paras$Basic$filter_sc$min.percentCells.event,
  min.Cell = paras$Basic$filter_sc$minCell,
  min.RC = paras$Basic$filter_sc$minRC,
  min.nCount = paras$Basic$filter_sc$min.nCount,
  min.nFeatures = paras$Basic$filter_sc$min.nFeatures,
  max.percentMT = paras$Basic$filter_sc$max.percentMT,
  cell.select = NULL
)
}
\arguments{
\item{paras}{list readSCSESconfig(paras_file)
Default min.percentCells.gene, min.percentCells.event, min.Cell,min.RC,
min.nCount, min.nFeatures, max.percentMT from paras}

\item{min.percentCells.gene}{the minimum percentage of cells that
are required to have at least 1 reads for a gene}

\item{min.percentCells.event}{the minimum percentage of cells that
are required to have psi=0/psi=1/psi=NA for a event}

\item{min.Cell}{the minimum number of cells that are required to have
at least min.RC reads.}

\item{min.RC}{the minimum number of reads required on at least min.Cell
cells for a junction.}

\item{min.nCount}{the minimum number of counts detected in a cell.}

\item{min.nFeatures}{the minimum number of features detected in a cell.}

\item{max.percentMT}{the maximum percentage of all the counts belonging
to a subset of mitochondrial genes in a cell.}

\item{cell.select}{cells selected for analysis, should match the name
of single-cell bam file name excluding the .bam suffix.}
}
\value{
filtered data path
}
\description{
process raw data in work_path/rds and
save filtered data to work_path/rds_processed
}
